logs/structformer_run/wandb/run-20250811_191748-44mib4iz/files/requirements.txt
jaxlib==0.6.2
nvidia-nccl-cu12==2.27.7
msgpack==1.1.1
hf-xet==1.1.7
zipp==3.23.0
gitdb==4.0.12
tzdata==2025.2
nvidia-cuda-nvrtc-cu12==12.9.86
frozenlist==1.7.0
wandb==0.21.1
nvidia-nvjitlink-cu12==12.9.86
click==8.2.1
mdurl==0.1.2
optax==0.2.5
sympy==1.14.0
setuptools==59.6.0
aiohttp==3.12.15
chex==0.1.90
jax-cuda12-pjrt==0.6.2
dotenv==0.9.9
mpmath==1.3.0
jax==0.6.2
networkx==3.4.2
nvidia-cuda-nvcc-cu12==12.9.86
GitPython==3.1.45
pydantic==2.11.7
dill==0.3.8
idna==3.10
urllib3==2.5.0
transformers==4.55.0
pytz==2025.2
nvidia-cuda-runtime-cu12==12.9.79
nest-asyncio==1.6.0
datasets==4.0.0
tqdm==4.67.1
aiohappyeyeballs==2.6.1
safetensors==0.6.2
platformdirs==4.3.8
protobuf==6.31.1
fsspec==2025.3.0
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cudnn-cu12==9.12.0.46
simplejson==3.20.1
xxhash==3.5.0
nvidia-nvshmem-cu12==3.3.20
ml_dtypes==0.5.3
tokenizers==0.21.4
flax==0.10.7
opt_einsum==3.4.0
scipy==1.15.3
nvidia-cublas-cu12==12.9.1.4
pyarrow==21.0.0
orbax-checkpoint==0.11.21
multiprocess==0.70.16
multidict==6.6.4
python-dotenv==1.1.1
smmap==5.0.2
jax-cuda12-plugin==0.6.2
etils==1.13.0
toolz==1.0.0
packaging==25.0
numpy==1.26.4
attrs==25.3.0
nvidia-cufft-cu12==11.4.1.4
typing_extensions==4.14.1
regex==2025.7.34
annotated-types==0.7.0
treescope==0.1.10
sentry-sdk==2.34.1
filelock==3.18.0
six==1.17.0
async-timeout==5.0.1
yarl==1.20.1
PyYAML==6.0.2
markdown-it-py==4.0.0
importlib_resources==6.5.2
tensorstore==0.1.76
nvidia-cusparse-cu12==12.5.10.65
Pygments==2.19.2
rich==14.1.0
requests==2.32.4
pydantic_core==2.33.2
nvidia-cusolver-cu12==11.7.5.82
python-dateutil==2.9.0.post0
certifi==2025.8.3
charset-normalizer==3.4.3
pip==25.2
typing-inspection==0.4.1
absl-py==2.3.1
pandas==2.3.1
torch==2.8.0
propcache==0.3.2
MarkupSafe==3.0.2
aiofiles==24.1.0
humanize==4.12.3
Jinja2==3.1.6
aiosignal==1.4.0
huggingface-hub==0.34.4
